backend/pyproject.toml
[project]
name = "ai-learning-platform-backend"
version = "0.1.0"
description = "AI Learning Platform backend"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "alembic==1.16.4",
    "asyncpg==0.30.0",
    "fastapi[standard]==0.115.14",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "sqlalchemy[asyncio]==2.0.43",
    "python-jose==3.5.0",
    "bcrypt==4.3.0",
    "gunicorn>=23.0.0",
    "openai>=1.40.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120
lint.select = ["ALL"]
lint.ignore = [
    "D1",
    "D203",
    "D212",
    "INP001",
    "ISC001",
    "COM812",
    "CPY001",
    "RUF001",
    "RUF003",
    "ERA001",
    "TD003",
    "FIX002",
]
exclude = ["alembic/env.py"]
preview = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
exclude = ["alembic/env.py"]

[tool.pytest.ini_options]
pythonpath = ["."]
